+++ source/extensions/filters/http/common/utility.h
@@ -2,11 +2,9 @@
 
 #include <unordered_map>
 
-#include "envoy/runtime/runtime.h"
-
-#include "common/common/logger.h"
 #include "common/common/macros.h"
 
+#include "extensions/common/utility.h"
 #include "extensions/filters/http/well_known_names.h"
 
 namespace Envoy {
@@ -23,7 +21,11 @@ public:
   /**
    * Given a deprecated HTTP filter extension name, return the
    * canonical name. Any name not defined in the deprecated map is
-   * returned without modification.
+   * returned without modification. If deprecated extension names are
+   * disabled, throws EnvoyException.
+   *
+   * @return const std::string& canonical filter name
+   * @throw EnvoyException if deprecated names are disabled
    */
   static const std::string&
   canonicalFilterName(const std::string& name,
@@ -33,34 +35,8 @@ public:
       return name;
     }
 
-    std::string err = fmt::format(
-        "Using deprecated http filter name '{}' for '{}'. This name will be removed from Envoy "
-        "soon. Please see "
-        "https://www.envoyproxy.io/docs/envoy/latest/intro/deprecated for details.",
-        name, it->second);
-
-#ifdef ENVOY_DISABLE_DEPRECATED_FEATURES
-    bool warn_only = false;
-#else
-    bool warn_only = true;
-#endif
-
-    if (runtime && !runtime->snapshot().deprecatedFeatureEnabled(
-                       "envoy.deprecated_features.allow_historic_http_filter_names", true)) {
-      warn_only = false;
-    }
-
-    if (warn_only) {
-      ENVOY_LOG_MISC(warn, "{}", err);
-    } else {
-      const char fatal_error[] =
-          " If continued use of this filter name is absolutely necessary, see "
-          "https://www.envoyproxy.io/docs/envoy/latest/configuration/operations/runtime"
-          "#using-runtime-overrides-for-deprecated-features for how to apply a temporary and "
-          "highly discouraged override.";
-
-      throw EnvoyException(err + fatal_error);
-    }
+    Extensions::Common::Utility::ExtensionNameUtil::checkDeprecatedExtensionName(
+        "http filter", name, it->second, runtime);
 
     return it->second;
   }